Move ExaMetadata to separate package
