+++ pkg/extensionAPI/exaMetadata/ExaMetadata_test.go
@@ -1,4 +1,4 @@
-package extensionAPI
+package exaMetadata
 
 import (
 	"fmt"
